requirements.txt
replicate==0.32.1
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
librosa==0.10.2.post1
soundfile==0.12.1
pydub==0.25.1
requests==2.32.3
tqdm==4.66.4
audioread==3.0.1

